Archive/requirements.txt
grpcio
grpcio-tools
librosa
opus-fast-mosestokenizer
faster-whisper